review(claude-tools): fix system prompt, separate example
